Proceeding with a hackish version of live-update.
I have achieved what we had with previous version of live-update. We
redraw whole page to get new changes in the template. It's not fully
tested or even fully working yet, some important parts are hardcoded, I
haven't even tested evetns and helpers yet, but hey! This shit is
working. Changes in the templates are changing live.