Add framework to register patches.
Patches are meant to detect and neutralize the code that might break or
do funny things when re`eval`ed. Like events and helpers code which
register events again when re`eval`ed.

A patch is basically an object with two keys:
{
  detector: function(code),
  neutralizer: function(code, match) { }
}

Before `eval`ing the code, code is passed through each patch.
* Code first goes to the detector of a patch
* If detector returns a falsy value, neutralizer is not executed
* If detector returns a truthy value, the code and the return value of
  detector are given to the neutralizer
* neutralizer must return the neutral/safe code. It can do some
  side-effect operations and return the same code
* what is returned by neutralizer is passed to next patch
* the end result of patches is evaled

The detector and neutralizer can be either functions or strings. If they
are a string say X, the `Eval.patches[X].detector`/`neutralizer` is used
for processing the code. I have second thoughts about whether we should
have this "feature" or not. It seem to be a possibly useful feature in
future for some cases and might help in re-usability, but at the same
time I also have thoughts that it will not be useful and we shall remove
it for the sake of simplicity. Since it is not adding that much
complexity I am keeping it.

XXX: **Reminder** Remove this feature if it turns out to be not useful