Handle changes in iron-router routes.
Add 'postEval' hook for patches which is a function that gets executed
after the code is `eval`ed. We need this for iron-router.

For iron-router, when the routes file is changed, we change the route to
present route. This handles the case when user changes the present route
path or params or anything pretty well. It also covers router
configuration, layout template change etc.

Eval.patches[N] now looks like this:
{
  detector: func,
  neutralizer: func,
  postEval: func
}

postEval is optional and is not passed any arguments.

Also added new signature for registerPatch. It can be given a string for
name and 3 functions, or a string for name and an object with 3 (or 2)
keys as stated above.