Fix autoruns eval for single autorun in a file.
This is how the fix works:
  We receive the old content along with the code to be `eval`ed. From
  the old content we pick all the functions that were fed to
  Tracker.autorun. If we find some, we find the computation that this
  function belongs to, and stop the computation. So before evaling the
  file, all the reactive computations that were being previously run
  from the file are stopped, and new ones are started with each eval.

  For finding which computation belongs to the function, we keep a
  collection of all the computations and then iterate over all of them.

There is room for performance improvement here; we might not need to
keep all the computations. Computations which belong to template views
can be omitted hence reducing the number of iterations we do.

This is not complete yet. It works for the first autorun it finds in the
file. Fix it for finding all the autoruns in file/code.