Make LiveUpdate._usingAsLib default to true.
To use liveupdate as standalone app, now it will be a requirement to set
`LiveUpdate.usingAsLib(false)`. This change is made because when
one client changes because of remotely connected nucleus and itself
hasn't inititalized a nucleus window, LiveUpdate works on it again. This
is easily fixable though. I have planned to fix this when I update
standalone LiveUpdate to use new eval class instead of old POC
method. Although that might never happen.