Make LiveUpdate to be used as a library-only package
That means we it don't do anything by itself. Adding this package and not configuring anything will only stop reloads on file change. It's possible to allow reloads on file change by configuring LiveUpdate as

```js
LiveUpdate.configure({interceptReloads: false});
```

Idea is to make this package a bare minimum, so that other packages can use it for pushing new template html/js. But this package still has CSS collection stuff which don't really belong here. It's because of its nature of being derived from Nucleus. It shall be moved to another package.

My plan is to create adaptors which will use this package. Live updating html/js in meteor AFTER meteor is done with rebuild process is pointless because that's what takes most time. The update should be instantaneous. For that, new js/html shall be provided to this package by hand. This is not a problem when done in Nucleus since all the new code is being editted in the app itself, but when using an external browser, there need to be a middle adaptor that picks new content and give it to this package. My next goal is to get the live-update working with Emacs, after that I might build an editor agnostic adaptor which would hot-push code on change in fs.