Make it work when iron:router is not present with a condition
Condition is that the html <body> code must be put in a main.html file. We need it because that's the base content which renders all other templates. Without it, LiveUpdate don't have anything to render new HTML into. Now we can make users pass <body> as an argument to `LiveUpdate.pushHtml`, but...later may be.